Get rid of extra slash in non-file titles.
Previously the window title looked like:
  /[No Name]
Now it is simply:
  [No Name]

I used a search and replace filename-modifier to suffix a trailing slash
if it matches an end of string. It won't produce the trailing slash if
there's no path available. The path and file name are two different
sections, the path generated from expand's %, and the filename from the
statusline variable %t. If the :h modifier results in an empty string,
the next modifier is ignored (I believe). There is some issues with
either escaping or restricted characters that mean I can only do certain
regexes here anyway, and I haven't been able to figure out whats allowed
or howto escape things that aren't. In any case, this particular
incantation behaves exactly as I intended thus far in my testing.