Bringing the readme up to date - 10 years!
Ten year anniversary! :tada:

This repo continues a long lineage of my configuration files, originally just scp'd between machines, then I think a brief stint in a Mercurial repo before finally migrating to git in 2011. 

Anyway, this cleans up a few things:

- I've switched from `ack` to `the-silver-search` and for now I've settled on `sift`  though I still include an `.ackrc` (for now)
- I never used the multiple-cursor thing in Vim, so I dropped it a long time ago.
- I also disabled the `vim-Lusty` suite after much tweaking of `CtrlP`.
- I ended up rarely using `Rails.vim` so I dropped it.

The whole repo has been reorganized a couple of times over the years but it's always kind of revolved around Vim. 
Last year I switched to `vim-Plug` which has made this repo much smaller and easier to manage, as well as making it trivial
to try out and remove plugins. Thanks so much to Junegunn Choi for creating it. 
I extracted most of the platform-agnostic setup code from this repo into Enrichen, but I haven't gotten around to fleshing it out to 
be used as a generic low-overhead installer yet. Someday! 

Anyway, here's to the next 10 years of configuration files.