neovim: intermediate but working state of a new config
Removed several unused plugins. Updated others to Lua equivalents.

Added BetterLua.vim, which makes Lua nicer to look at but I think messed
up my auto-tab settings somehow.

Added Obsidian.nvim, but I haven't messed with it yet.

All of this is subject to change but it works as-is.

- LuaLine

Replaces LightLight.vim. Written in Lua. Slightly nicer. Easy to
configure. Possibly faster. I've modified it to be very similar to my
LightLine setup, jusy shinier.

- Telescope

WIP replacing CtrlP. Nicer UI by far. Lots of plugins support it. But I
am considering ditching it in favor of nvim-fzf which comes with more
support built in and uses the real fzf for faster and more configurable
fuzzy finding.

- Incline

Used with LuaLine's global statusline. Meaning that the statusline at
the bottom is the one that always displays info instead of one
statusline for each window pane. Incline adds small labels with
filenames (and whatever information you want).

I'm not fully sold on them. It requires a lot of work to set it up.

I've spent many hours learning Lua and NeoVim's API to integrate it with
LuaLine and my colorscheme. Right now I have it mirroring LuaLine
theme when the editor changes mode. Now I'm trying to make that work
with active/inactive windows, so that it is easy to tell at a glance
which one is being edited. (all things that non-global statuslines do for
free!)