neovim: Incline now dynamically pulls theme from LuaLine
No longer needs to specify the name of the them manually. And now
updates as soon as the LueLine theme updates.

Also falls back to using the file type icon color if LuaLine is not
available.

Removed debug code that I was using to get my Lua-legs.
Refactor and general cleanup.